Shooting mechanic for players more or less finished
NOTE: Animations may make us tweak a few things about the mechanics

+++ BadranOmarCS4423Final/Assets/Enemy.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int health = 100;
+    public float health = 100f;
     //public GameObject enemy;
     
-    public void TakeDamage (int damage){
+    public void TakeDamage (float damage){
         health -= damage;
 
         if(health<=0){
